a lot of changes to resolve issues with rotation & reload pre-iOS 6.0
xolawareWebViewXoom is intended to solve the problem of allowing a
vertical only zoom (only allowed if auto-scale is turned off) and
vertical only scrolling (only allowed if auto-scale is turned on.

and the pre-iOS 6.0 version of UIWebView auto-scaling does not deal well
with having this setting turned on and off.

the manifestation that is the worst problem is that the text in the
bottom of a webView will be cut-off after a javaScript setMaxWidth has
been performed on zoomed text after auto-zooming has completed following
a rotation or the reloading of a page after an anchor has been removed.

there are so many changes that this commit comment is going to refer to
the code comments that have been added to describe the host of changes
that have been made.