iOS tautology … minimumScaleFactor/minimumFontSize usage
when __IPHONE_OS_VERSION_MIN_REQUIRED is less than __IPHONE_6_0, IB
won't compile .storyboard if it contains minimumScaleFactor.

thus, checking respondsToSelector:@selector(minimumScaleFactor) is
meaningless since there won't be a need to check minimumScaleFactor when
running a build created with a target less than 6.0, and it will always
be there when building a target based 6.0 or above.

instead, just do one for builds targeted at 6.0 and above, and the other
when targeting older builds.